Refactor/cleanup some string writers
