+++ strings/base_module_ixx_preamble.h
@@ -1,0 +1,9 @@
+module;
+#define WINRT_IMPL_BUILD_MODULE
+#include <intrin.h>
+#include <cstddef>
+#include <version>
+#ifdef _DEBUG
+#include <crtdbg.h>
+#endif // _DEBUG
+#include "winrt/module.h"
